requirements.txt
asn1crypto==0.24.0
blinker==1.4
certifi==2018.8.13
cffi==1.11.5
chardet==3.0.4
configobj==5.0.6
cryptography==2.3.1
defusedxml==0.5.0
django-bootstrap-themes==3.3.6
django-bootstrap3==10.0.1
django-debug-toolbar==1.9.1
django-registration-redux==2.4
django-tastypie==0.14.1
httplib2==0.9.2
idna==2.7
Jinja2==2.10
jsonpatch==1.16
jsonpointer==1.10
jsonschema==2.6.0
keyring==10.6.0
keyrings.alt==3.0
MarkupSafe==1.0
oauthlib==2.1.0
paypalrestsdk==1.13.1
Pillow==5.2.0
python-social-auth==0.3.6
python3-openid==3.1.0
pytz==2018.5
pyxdg==0.25
PyYAML==3.12
requests==2.19.1
requests-oauthlib==1.0.0
requests-unixsocket==0.1.5
SecretStorage==2.3.1
six==1.11.0
social-auth-app-django==2.1.0
social-auth-core==1.7.0
sqlparse==0.2.4
stripe==2.3.0
urllib3==1.23
uWSGI==2.0.17.1
